categorize windows.processghosting as a malware plugin
